+++ tests/unit/controllers/PurseControllerTest.php
@@ -10,7 +10,17 @@
 
 namespace hipanel\modules\finance\tests\unit\controllers;
 
+use hipanel\modules\finance\actions\GenerateAndSaveDocumentAction;
 use hipanel\modules\finance\controllers\PurseController;
+use hipanel\modules\finance\tests\unit\support\TestHiartResponse;
+use hipanel\modules\finance\tests\unit\support\TestMessageSource;
+use hipanel\modules\finance\tests\unit\support\TestRequest;
+use hipanel\modules\finance\tests\unit\support\TestSession;
+use hipanel\modules\finance\tests\unit\support\TestUser;
+use hiqdev\hiart\ResponseErrorException;
+use Yii;
+use yii\base\Component;
+use yii\web\Response;
 
 /**
  * Generated by PHPUnit_SkeletonGenerator on 2016-04-27 at 13:36:04.
@@ -19,17 +29,110 @@ class PurseControllerTest extends \PHPUnit\Framework\TestCase
 {
     protected PurseController $object;
 
+    private object $originalSession;
+    private object $originalUser;
+    private array $originalI18nTranslations;
+
     protected function setUp(): void
     {
         $this->object = new PurseController('test', null);
+
+        $this->originalSession = Yii::$app->session;
+        $this->originalUser = Yii::$app->user;
+        $this->originalI18nTranslations = Yii::$app->i18n->translations;
+
+        Yii::setAlias('@requisite', '/requisite');
+        Yii::setAlias('@client', '/client');
+        Yii::$app->set('session', new TestSession());
+        Yii::$app->set('user', new TestUser(['canMap' => ['requisites.update' => true]]));
+        Yii::$app->i18n->translations['hipanel*'] = ['class' => TestMessageSource::class];
     }
 
     protected function tearDown(): void
     {
+        Yii::$app->set('session', $this->originalSession);
+        Yii::$app->set('user', $this->originalUser);
+        Yii::$app->i18n->translations = $this->originalI18nTranslations;
     }
 
     public function testActions(): void
     {
-        $this->assertIsArray($this->object->actions());
+        $actions = $this->object->actions();
+
+        $this->assertIsArray($actions);
+        $this->assertSame(GenerateAndSaveDocumentAction::class, $actions['generate-and-save-monthly-document']['class']);
+        $this->assertSame(GenerateAndSaveDocumentAction::class, $actions['generate-acts']['class']);
+        $this->assertSame(GenerateAndSaveDocumentAction::class, $actions['generate-and-save-document']['class']);
+    }
+
+    public function testGenerateDocumentShowsGenericErrorWhenErrorOpsHasUnrecognizedErrorMessage(): void
+    {
+        $payload = [
+            '_error' => 'failed validate document payload',
+            '_error_ops' => [
+                'requisite_id' => 42,
+                'type' => 'purchase_invoice',
+            ],
+        ];
+
+        $controller = $this->createControllerThrowing($this->createResponseErrorException($payload));
+        $controller->generateDocument('generate-document', ['client_id' => 1]);
+
+        $session = Yii::$app->session;
+        $this->assertInstanceOf(TestSession::class, $session);
+        $flash = $session->flashes['error'][0] ?? null;
+        $this->assertSame('Failed to generate document', $flash);
+    }
+
+    public function testGenerateDocumentShowsRequisiteLinkWhenErrorOpsHasTemplateError(): void
+    {
+        $requisiteId = 99;
+        $payload = [
+            '_error' => 'failed find document template',
+            '_error_ops' => [
+                'requisite_id' => $requisiteId,
+                'type' => 'purchase_invoice',
+            ],
+        ];
+
+        $controller = $this->createControllerThrowing($this->createResponseErrorException($payload));
+        $controller->generateDocument('generate-document', ['client_id' => 1]);
+
+        $session = Yii::$app->session;
+        $this->assertInstanceOf(TestSession::class, $session);
+        $flash = $session->flashes['error'][0] ?? null;
+        $this->assertIsString($flash);
+        $this->assertStringContainsString('No templates for requisite', $flash);
+        $this->assertStringContainsString((string)$requisiteId, $flash);
+    }
+
+    private function createControllerThrowing(ResponseErrorException $e): PurseController
+    {
+        return new class('test', null, $e) extends PurseController {
+            public function __construct(
+                string $id,
+                mixed $module,
+                private readonly ResponseErrorException $exception,
+                array $config = [],
+            ) {
+                parent::__construct($id, $module, $config);
+            }
+
+            protected function performDocumentAction(string $action, array $params): mixed
+            {
+                throw $this->exception;
+            }
+
+            public function redirect($url, $statusCode = 302): Response
+            {
+                return new Response();
+            }
+        };
+    }
+
+    private function createResponseErrorException(array $payload): ResponseErrorException
+    {
+        $response = new TestHiartResponse(new TestRequest(), $payload);
+        return new ResponseErrorException('error', $response);
     }
 }
