+++ tests/unit/support/TestMessageSource.php
@@ -1,0 +1,14 @@
+<?php
+declare(strict_types=1);
+
+namespace hipanel\modules\finance\tests\unit\support;
+
+use yii\i18n\MessageSource;
+
+final class TestMessageSource extends MessageSource
+{
+    protected function translateMessage($category, $message, $language): string
+    {
+        return $message;
+    }
+}
